DataIngestion/moea_optimizer/pyproject.toml
[tool.poetry]
name = "moea-optimizer"
version = "1.0.0"
description = "Multi-Objective Evolutionary Algorithm optimizer for greenhouse climate control"
authors = ["Greenhouse AI Team"]
readme = "README.md"
python = "^3.10,<3.12"  # PyTorch compatibility constraint
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
# CPU MOEA
pymoo = "^0.6.1.3"
# GPU MOEA & Deep Learning
evox = "^1.3.0"
torch = {version = "^2.7.0", source = "pytorch-cu118"}
# Data & Stats
numpy = "^2.0.0"
pandas = "^2.2.3"
scipy = "^1.15.0"
scikit-posthocs = "^0.11.4"
# Database & Config
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
toml = "^0.10.2"
pyyaml = "^6.0"
# Visualization
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
plotly = "^5.0.0"
# Utilities
click = "^8.0.0"
tqdm = "^4.0.0"
joblib = "^1.3.0"
# Monitoring
tensorboard = "^2.0.0"
wandb = "^0.16.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.0.0"
pre-commit = "^3.0.0"
ruff = "^0.1.0"

[tool.poetry.group.notebooks.dependencies]
jupyter = "^1.0.0"
jupyterlab = "^4.0.0"
ipywidgets = "^8.0.0"

[[tool.poetry.source]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "DTZ", "RUF"]
ignore = [
    "E501",  # Line length handled by formatter
    "N803",  # Argument name should be lowercase (F, X are standard in optimization)
    "N806",  # Variable in function should be lowercase (F, X are standard)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.poetry.scripts]
moea-optimizer = "src.cli:main"
moea-benchmark = "src.benchmark:main"